Python3/googledrive/requirements.txt
google-api-python-client==1.7.3
oauth2client==4.1.2